Disable networking, enable AppArmor profile
For especially hardened deployments, force all containers
started by ECS agent to have no access to network and be
restricted by an AppArmor profile.

Note that because go-dockerclient is version-pinned with
Godeps, introducing the specific change to add support for
the `SecurityOpts` option took place in the Coursera fork
of go-dockerclient.